move writeOutput into deploy helpers
